requirements.txt
asn1crypto==1.5.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
defusedxml==0.7.1
endesive==2.19.3
Flask==3.1.2
fonttools==4.61.1
fpdf2==2.8.5
idna==3.11
invoke==2.2.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.2
pandas==3.0.0
paramiko==4.0.0
pillow==12.1.0
platformdirs==4.5.1
pycparser==3.0
PyKCS11==1.5.18
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
zeep==4.3.2
